% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armgin-package.R
\docType{package}
\name{armgin-package}
\alias{armgin}
\alias{armgin-package}
\title{Un-pivot complex and irregular data layouts.}
\description{
The armgin package summarises all combinations of grouping variables at
once. For example, vote counts by
\itemize{
\item voting place, ward, constituency, party, candidate
\item voting place, ward, constituency, party
\item voting place, ward, constituency
\item voting place, ward,
\item voting place
}

Or with \code{hierarchy = FALSE} and grouping by only voting place, ward and
constituency (for brevity)
\itemize{
\item voting place, ward, constituency
\item voting place, ward
\item voting place, constituency
\item ward, constituency
\item voting place
\item ward
\item constituency
}

Groups are defined as normal by \code{\link[dplyr:group_by]{dplyr::group_by()}} and then piped into
\code{\link[=margins]{margins()}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nacnudus/armgin}
  \item Report bugs at \url{https://github.com/nacnudus/armgin/issues}
}

}
\author{
\strong{Maintainer}: Duncan Garmonsway \email{nacnudus@gmail.com}

}
\keyword{internal}
