% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{Summarise all margins of a grouped data frame in one go}
\usage{
margins(.data, ..., bind = TRUE, hierarchy = TRUE)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables to group by. All tbls accept variable names.
Some tbls will accept functions of variables. Duplicated groups
will be silently dropped.}

\item{bind}{Whether to combine all margins into one data frame. Default is
\code{TRUE}.  If \code{FALSE}, returns a list of dataframes, one per margin.}

\item{hierarchy}{Whether to treat the grouping variables as a hierarchy with
the first variable at the top and the next variable inside it.  Default is
\code{TRUE}.  If \code{FALSE}, uses all unordered combinations of the grouping
variables as margins.}
}
\description{
Take a data frame grouped with [dplyr::group_by()) and summarise by
combinations of the grouping variables, e.g. by the first variable, by the
first two, the first three, etc.  With \code{hierarchy = FALSE} all combinations
will be summarised, e.g. \code{a}, \code{a} and \code{b}, \code{a} and \code{c}, \code{b} and \code{c}.

The output is a data frame with \code{NA} in unused grouping variables.
Alternatively use \code{bind = FALSE} to return a list of data frames, one per
combination of grouping variables.
}
\examples{
mtcars \%>\%
  dplyr::group_by(cyl, gear, am) \%>\%
  margins(mpg = mean(mpg, na.rm = TRUE),
          hp = min(hp),
          bind = FALSE,
          hierarchy = TRUE)
mtcars \%>\%
  dplyr::group_by(cyl, gear, am) \%>\%
  margins(mpg = mean(mpg, na.rm = TRUE),
          hp = min(hp),
          bind = TRUE,
          hierarchy = TRUE) \%>\%
  print(n = Inf)
mtcars \%>\%
  dplyr::group_by(cyl, gear, am) \%>\%
  margins(mpg = mean(mpg, na.rm = TRUE),
          hp = min(hp),
          bind = FALSE,
          hierarchy = FALSE)
}
